LayoutTests/fast/speech/scripted/speechrecognition-errors-expected.txt

Test Speech JavaScript API errors

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitSpeechRecognition' in self is true
PASS webkitSpeechRecognition == null is false

notAllowedTest():
onerror
PASS count is 0
PASS event.code is webkitSpeechRecognitionError.NOT_ALLOWED
PASS event.message is "not allowed"
PASS event.type is "error"
onend
PASS count is 1
PASS successfullyParsed is true

TEST COMPLETE
